daxctl: link against libndctl, in case its use doesn't get pruned
util/json.c uses libndctl symbols, and is included by daxctl.  These
functions should then get pruned as unused, but on some platforms the
toolchain fails to do so.

These platforms are ia64, hppa and 32-bit powerpc.  It's generally a
waste of our time to build ndctl there, but as the lack of a
build-dependency requires annoying workarounds higher in the stack,
this has been requested in https://bugs.debian.org/914348 -- and is
trivially fixable on the ndctl side.

Thanks to -Wl,--as-needed there's no harm to architectures where the
pruning works, thus let's not bother with detection.

As daxctl and libdaxctl are separate, there's no circular dependency.

Signed-off-by: Adam Borowski <kilobyte@angband.pl>