test/cxl-topology.sh: change assumption on host bridge validation
Current host bridge validation in cxl-topology.sh assumes that the
decoder enumeration is in order and therefore the port numbers can
be used as a sorting key. With delayed port enumeration, this
assumption is no longer true. Change the sorting to by number
of children ports for each host bridge as the test code expects
the first 2 host bridges to have 2 children and the third to only
have 1.

Signed-off-by: Dave Jiang <dave.jiang@intel.com>
Signed-off-by: Vishal Verma <vishal.l.verma@intel.com>
Reviewed-by: Alison Schofield <alison.schofield@intel.com>
Tested-by: Li Zhijian <lizhijian@fujitsu.com>
Link: https://lore.kernel.org/r/20250508204419.3227297-1-dave.jiang@intel.com
Signed-off-by: Alison Schofield <alison.schofield@intel.com>