cxl/test: skip, don't fail, when kernel tracing is not enabled
cxl-events.sh and cxl-poison.sh require a kernel with CONFIG_TRACING
enabled and currently report a FAIL when /sys/kernel/tracing is
missing. Update these tests to report a SKIP along with a message
stating the requirement. This allows the tests to run cleanly on
systems without TRACING enabled and gives users the info needed to
enable TRACING for testing.

Reviewed-by: Dave Jiang <dave.jiang@intel.com>
Reviewed-by: Marc Herbert <marc.herbert@linux.intel.com>
Reviewed-by: Dan Williams <dan.j.williams@intel.com>
Reported-by: Itaru Kitayama <itaru.kitayama@linux.dev>
Closes: http://lore.kernel.org/FD4183E1-162E-4790-B865-E50F20249A74@linux.dev
Link: https://lore.kernel.org/r/20250523031518.1781309-1-alison.schofield@intel.com
Signed-off-by: Alison Schofield <alison.schofield@intel.com>